+++ src/network/player/Player.hpp
@@ -23,8 +23,8 @@ public:
     void SetId(uint8_t id) noexcept { id_ = id; }
     [[nodiscard]] uint8_t GetId() const noexcept { return id_; }
 
-    void SetCharacterId(uint8_t id) noexcept { character_id_ = id; }
-    [[nodiscard]] uint8_t GetCharacterId() const noexcept { return character_id_; }
+    void SetCharacterId(uint16_t id) noexcept { character_id_ = id; }
+    [[nodiscard]] uint16_t GetCharacterId() const noexcept { return character_id_; }
 
     void SetNetInfo(ClientInfo* net_info) noexcept { net_info_ = net_info; }
     [[nodiscard]] ClientInfo* GetNetInfo() const noexcept { return net_info_; }
@@ -34,7 +34,7 @@ public:
 
 private:
     uint8_t id_{ 0 };
-    uint8_t character_id_{ 0 };
+    uint16_t character_id_{ 0 };
     ClientInfo* net_info_{ nullptr };
 };
